requirements/requirements.txt
aioredis==2.0.1
amqp==5.2.0
anyio==4.6.2
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==5.1.2
django-celery-beat==2.7.0
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-pydantic-field==0.3.10
django-templated-mail==1.1.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docker==7.1.0
flower==2.0.1
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httptools==0.6.4
humanize==4.11.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
model-bakery==1.20.0
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.1
Pillow==11.0.0
prometheus-client==0.21.0
prompt-toolkit==3.0.48
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
pydantic==2.9.2
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
sentry-sdk==2.17.0
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tornado==6.4.1
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
wrapt==1.16.0
zope.interface==7.1.1
